moved GetMethodType function to generator_helpers.h
